test : 검증부분 별도 클래스로 분리
